put interactive and dry check into the inwx_api() function
